requirements.txt
# Reddit API
praw>=7.7.0
requests>=2.31

# Data handling
pandas>=2.0

# Configuration
python-dotenv>=1.0

# Logging
loguru>=0.7

# NLP - Análisis de sentimiento
transformers>=4.40
torch>=2.0
vaderSentiment>=3.3.2

# NLP - Modelado de tópicos
bertopic>=0.16
sentence-transformers>=2.7
scikit-learn>=1.3

# Coherencia temática (evaluación)
gensim>=4.3
